+++ app/src/main/java/com/bosseurs/medcare/ui/screens/obesite/obesiteState.kt
@@ -1,8 +1,8 @@
 package com.bosseurs.medcare.ui.screens.obesite
 
 data class obesiteState(
-val taille: Float = 0f,
-val poids : Float = 0f ,
+val taille: Int = 0 ,
+val poids : Int = 0 ,
 val genre : Boolean = false
 
 )
